bugfix: account type => long, added properties

+++ src/main/java/com/kloudless/model/Event.java
@@ -4,8 +4,8 @@ import com.kloudless.net.APIResourceMixin;
 
 public class Event extends APIResourceMixin {
 	public String id;
-	public Integer event_id;
-	public Integer account;
+	public Long event_id;
+	public Long account;
 	public String action;
 	public String modified;
 	public String type;
